Consolidate release image utility and relocate to deployment utils
